requirements.txt
discord.py>=2.3.2
motor>=3.1.1
aiohttp
python-dotenv